requirements.txt
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
bcrypt==4.0.0
beanie==1.10.9
billiard==3.6.4.0
celery==5.2.7
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.78.0
filelock==3.8.0
gevent==21.12.0
greenlet==1.1.3
h11==0.14.0
identify==2.5.5
idna==3.4
kombu==5.2.4
marshmallow==3.18.0
motor==2.5.1
multidict==6.0.2
nodeenv==1.7.0
packaging==21.3
passlib==1.7.4
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.31
pydantic==1.9.2
PyJWT==2.4.0
pymongo==3.12.3
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
umongo==3.1.0
uvicorn==0.17.6
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.4.0
